pyproject.toml
[tool.poetry]
name = "psychologyservice"
version = "0.1.0"
description = ""
authors = ["ILya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
greenlet = "^3.1.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
